Moves scales to different files
