+++ web/src/features/exchanges/ExchangeArrow.tsx
@@ -4,8 +4,8 @@ import { useMemo } from 'react';
 import { MapboxMap } from 'react-map-gl';
 import { resolvePath } from 'react-router-dom';
 import { ExchangeArrowData } from 'types';
-import { quantizedCo2IntensityScale, quantizedExchangeSpeedScale } from 'utils/scales';
 import ExchangeTooltip from './ExchangeTooltip';
+import { quantizedCo2IntensityScale, quantizedExchangeSpeedScale } from './scales';
 
 interface ExchangeArrowProps {
   data: ExchangeArrowData;
